nowLongForSales

+++ src/main/java/ru/netology/stats/Main.java
@@ -5,7 +5,7 @@ import java.util.Arrays;
 public class Main {
     public static void main(String[] args) {
         StatisticsBiseness service = new StatisticsBiseness();
-        int[] arr = {
+        long[] arr = {
                 8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18
         };
         System.out.println();
