+++ src/main/java/ru/netology/stats/StatisticsBiseness.java
@@ -2,20 +2,20 @@ package ru.netology.stats;
 
 public class StatisticsBiseness {
 
-    public int calculateSumSales(int[] sales) {
-        int sum = 0;
-        for (int sale : sales) {
+    public long calculateSumSales(long[] sales) {
+        long sum = 0;
+        for (long sale : sales) {
             sum += sale;
         }
         return sum;
     }
 
-    public int findAverage(int[] sales) {
-        int sumSales = calculateSumSales(sales);
+    public long findAverage(long[] sales) {
+        long sumSales = calculateSumSales(sales);
         return calculateSumSales(sales) / sales.length;
     }
 
-    public int findNumberMonthWithMaxSales(int[] sales) {
+    public long findNumberMonthWithMaxSales(long[] sales) {
         int monthMaxSales = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] >= sales[monthMaxSales]) {
@@ -26,7 +26,7 @@ public class StatisticsBiseness {
         return monthMaxSales + 1;
     }
 
-    public int findNumberMonthWithMinSales(int[] sales) {
+    public long findNumberMonthWithMinSales(long[] sales) {
         int monthMinSales = 0;
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] <= sales[monthMinSales]) {
@@ -36,10 +36,10 @@ public class StatisticsBiseness {
         return monthMinSales + 1;
     }
 
-    public int findMonthsWithSalesBelowAverage(int[] sales) {
-        int monthsBelowAverage = 0;
-        int average = calculateSumSales(sales) / sales.length;
-        for (int i : sales) {
+    public long findMonthsWithSalesBelowAverage(long[] sales) {
+        long monthsBelowAverage = 0;
+        long average = calculateSumSales(sales) / sales.length;
+        for (long i : sales) {
             if (i < average) {
                 monthsBelowAverage++;
             }
@@ -47,10 +47,10 @@ public class StatisticsBiseness {
         return monthsBelowAverage;
     }
 
-    public int findMonthsWithSalesHigherAverage(int[] sales) {
-        int monthsHigherAverage = 0;
-        int average = calculateSumSales(sales) / sales.length;
-        for (int i : sales) {
+    public long findMonthsWithSalesHigherAverage(long[] sales) {
+        long monthsHigherAverage = 0;
+        long average = calculateSumSales(sales) / sales.length;
+        for (long i : sales) {
             if (i > average) {
                 monthsHigherAverage++;
             }
